WIP: refactoring callees
