+++ core/src/constants/COLLATERALS.ts
@@ -6,8 +6,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'AAVE-A',
         symbol: 'AAVE',
         decimals: 18,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -16,8 +16,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'BAL-A',
         symbol: 'BAL',
         decimals: 18,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -26,8 +26,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'BAT-A',
         symbol: 'BAT',
         decimals: 18,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -36,8 +36,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'COMP-A',
         symbol: 'COMP',
         decimals: 18,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -46,8 +46,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'ETH-A',
         symbol: 'ETH',
         decimals: 18,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: [],
         },
     },
@@ -56,8 +56,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'ETH-B',
         symbol: 'ETH',
         decimals: 18,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: [],
         },
     },
@@ -66,8 +66,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'ETH-C',
         symbol: 'ETH',
         decimals: 18,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: [],
         },
     },
@@ -76,8 +76,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'GUSD-A',
         symbol: 'GUSD',
         decimals: 2,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -86,8 +86,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'KNC-A',
         symbol: 'KNC',
         decimals: 18,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -96,8 +96,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'LINK-A',
         symbol: 'LINK',
         decimals: 18,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -106,8 +106,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'LRC-A',
         symbol: 'LRC',
         decimals: 18,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -116,8 +116,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'MANA-A',
         symbol: 'MANA',
         decimals: 18,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -126,8 +126,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'PAXUSD-A',
         symbol: 'PAXUSD',
         decimals: 18,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -136,8 +136,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'RENBTC-A',
         symbol: 'RENBTC',
         decimals: 8,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -146,8 +146,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'TUSD-A',
         symbol: 'TUSD',
         decimals: 18,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -156,8 +156,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'UNI-A',
         symbol: 'UNI',
         decimals: 18,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -166,8 +166,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'USDC-A',
         symbol: 'USDC',
         decimals: 6,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -176,8 +176,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'USDC-B',
         symbol: 'USDC',
         decimals: 6,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -186,8 +186,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'USDT-A',
         symbol: 'USDT',
         decimals: 6,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -196,8 +196,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'WBTC-A',
         symbol: 'WBTC',
         decimals: 8,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -206,8 +206,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'YFI-A',
         symbol: 'YFI',
         decimals: 18,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -216,8 +216,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'ZRX-A',
         symbol: 'ZRX',
         decimals: 18,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -226,8 +226,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'MATIC-A',
         symbol: 'MATIC',
         decimals: 18,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -236,8 +236,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'WBTC-B',
         symbol: 'WBTC',
         decimals: 8,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -246,8 +246,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'WBTC-C',
         symbol: 'WBTC',
         decimals: 8,
-        uniswap: {
-            type: 'token',
+        exchange: {
+            callee: 'UniswapV2CalleeDai',
             route: ['ETH'],
         },
     },
@@ -256,8 +256,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'UNIV2DAIETH-A',
         symbol: 'UNIV2DAIETH',
         decimals: 18,
-        uniswap: {
-            type: 'lpToken',
+        exchange: {
+            callee: 'UniswapV2LpTokenCalleeDai',
             token0: 'DAI',
             token1: 'ETH',
         },
@@ -267,8 +267,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'UNIV2USDCETH-A',
         symbol: 'UNIV2USDCETH',
         decimals: 18,
-        uniswap: {
-            type: 'lpToken',
+        exchange: {
+            callee: 'UniswapV2LpTokenCalleeDai',
             token0: 'USDC',
             token1: 'ETH',
         },
@@ -278,8 +278,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'UNIV2ETHUSDT-A',
         symbol: 'UNIV2ETHUSDT',
         decimals: 18,
-        uniswap: {
-            type: 'lpToken',
+        exchange: {
+            callee: 'UniswapV2LpTokenCalleeDai',
             token0: 'ETH',
             token1: 'USDT',
         },
@@ -289,8 +289,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'UNIV2WBTCDAI-A',
         symbol: 'UNIV2WBTCDAI',
         decimals: 18,
-        uniswap: {
-            type: 'lpToken',
+        exchange: {
+            callee: 'UniswapV2LpTokenCalleeDai',
             token0: 'WBTC',
             token1: 'DAI',
         },
@@ -300,8 +300,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'UNIV2WBTCETH-A',
         symbol: 'UNIV2WBTCETH',
         decimals: 18,
-        uniswap: {
-            type: 'lpToken',
+        exchange: {
+            callee: 'UniswapV2LpTokenCalleeDai',
             token0: 'WBTC',
             token1: 'ETH',
         },
@@ -311,8 +311,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'UNIV2LINKETH-A',
         symbol: 'UNIV2LINKETH',
         decimals: 18,
-        uniswap: {
-            type: 'lpToken',
+        exchange: {
+            callee: 'UniswapV2LpTokenCalleeDai',
             token0: 'LINK',
             token1: 'ETH',
         },
@@ -322,8 +322,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'UNIV2UNIETH-A',
         symbol: 'UNIV2UNIETH',
         decimals: 18,
-        uniswap: {
-            type: 'lpToken',
+        exchange: {
+            callee: 'UniswapV2LpTokenCalleeDai',
             token0: 'UNI',
             token1: 'ETH',
         },
@@ -333,8 +333,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'UNIV2AAVEETH-A',
         symbol: 'UNIV2AAVEETH',
         decimals: 18,
-        uniswap: {
-            type: 'lpToken',
+        exchange: {
+            callee: 'UniswapV2LpTokenCalleeDai',
             token0: 'AAVE',
             token1: 'ETH',
         },
@@ -344,8 +344,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'UNIV2DAIUSDT-A',
         symbol: 'UNIV2DAIUSDT',
         decimals: 18,
-        uniswap: {
-            type: 'lpToken',
+        exchange: {
+            callee: 'UniswapV2LpTokenCalleeDai',
             token0: 'DAI',
             token1: 'USDT',
         },
@@ -355,8 +355,8 @@ const COLLATERALS: Record<string, CollateralConfig> = {
         ilk: 'UNIV2DAIUSDC-A',
         symbol: 'UNIV2DAIUSDC',
         decimals: 18,
-        uniswap: {
-            type: 'lpToken',
+        exchange: {
+            callee: 'UniswapV2LpTokenCalleeDai',
             token0: 'DAI',
             token1: 'USDC',
         },
