+++ core/src/types.ts
@@ -56,13 +56,13 @@ export declare interface AuctionTransaction extends Auction, TransactionFees {
     transactionProfitMinusFees: BigNumber;
 }
 
-export declare interface UniswapTokenConfig {
-    type: 'token';
+export declare interface UniswapV2CalleeConfig {
+    callee: 'UniswapV2CalleeDai';
     route: string[];
 }
 
-export declare interface UniswapLpTokenConfig {
-    type: 'lpToken';
+export declare interface UniswapV2LpTokenCalleeConfig {
+    callee: 'UniswapV2LpTokenCalleeDai';
     token0: string;
     token1: string;
 }
@@ -72,7 +72,7 @@ export declare interface CollateralConfig {
     ilk: string;
     symbol: string;
     decimals: number;
-    uniswap: UniswapTokenConfig | UniswapLpTokenConfig;
+    exchange: UniswapV2CalleeConfig | UniswapV2LpTokenCalleeConfig;
 }
 
 export declare interface NetworkConfig {
@@ -81,8 +81,18 @@ export declare interface NetworkConfig {
     url: string;
     gasPrice?: number;
     etherscanUrl: string;
-    uniswapV2CalleeDaiAddress: string;
-    uniswapV2LpTokenCalleeDaiAddress: string;
+}
+
+export declare interface CalleeAddresses {
+    UniswapV2CalleeDai: string;
+    UniswapV2LpTokenCalleeDai: string;
+}
+
+export type CalleeNames = keyof CalleeAddresses;
+
+export declare interface CalleeFunctions {
+    getCalleeData: (network: string, collateral: CollateralConfig, profitAddress: string) => Promise<string>;
+    getMarketPrice: (network: string, collateral: CollateralConfig, amount: BigNumber) => Promise<BigNumber>;
 }
 
 export declare interface MakerParams {
