+++ frontend/store/auctions.ts
@@ -7,7 +7,6 @@ import {
     restartAuction,
     enrichAuctionWithPriceDropAndMarketValue,
 } from 'auctions-core/src/auctions';
-import { checkAllExchangeRates } from 'auctions-core/src/uniswap';
 import { enrichAuctionWithTransactionFees } from 'auctions-core/src/fees';
 import { checkAllCalcParameters } from 'auctions-core/src/params';
 import { checkAllSupportedCollaterals } from 'auctions-core/src/addresses';
@@ -218,10 +217,6 @@ export const actions = {
                 error instanceof Error && error.message
             );
         }
-    },
-    async checkAllExchangeRates({ rootGetters }: ActionContext<State, State>) {
-        const network = rootGetters['network/getMakerNetwork'];
-        await checkAllExchangeRates(network);
     },
     async checkAllCalcParameters({ rootGetters }: ActionContext<State, State>) {
         const network = rootGetters['network/getMakerNetwork'];
